refactor: attack systeme so now it shoots regularly depending on the firerate component
